fixed : i2c speed for OLED Driver

+++ config.h
@@ -8,4 +8,4 @@
 
 #define NULL ((void*)0)
 
-#define I2C_CLK_FREQ_HZ 100000
+#define I2C_CLK_FREQ_HZ (int)100000
